Add complex restriction check.  (#479)
* Fix the separation of transition costs in the reverse path. Improperly
added entire cost, rather than separating out the elapsed_time portion
so the reverse path can be properly constructed with elapsed time to
the end of each edge.

* Add support for via-way (complex) restrictions.

* Disable destination only for bidirectional A*.

* Add complex restriction check. Fixes "trivial" path test where destination
is on same edge as origin but in opposing direction of oneway travel. This
case falls back to A* logic and needs to honor complex restrictions.

* Protect against crashes that can occur if an edge is shorter than the lat,lng
tolerance.