Use actual distance along the shape (#498)
* Set bounding box for trivial path.

* aAdd check for complex restrictions.

* Fix typo in comment.

* Check for complex restrictions

* Use actual distance along shape when rejecting edges (rather than distance
from prior node to end node). This allows rejection of edges that might be
a U-turn vs. the shape which goes around the block and never exceeds the
threshold distance. For efficiency it computes lengths between shape points
once and reuses those when computing lengths along portions of the shape.