Interrupts - prepare for adding to other thor service calls (#527)
* Pass interrupt method into TripPathBuilder::Build. If not nullptr will
call the interrupt callback method supplied.

* Add interrupt_callback data member and set it to nullptr in the constructor. The
callback method gets set in the work method. This will allow use of the interrupt
in other thor service methods.

* Set the interrupt callback when calling TripPathBuilder::Build

* Use a simple integer that is incremented with each iteration of the loop. This
is simpler than the prior approach to checking for interrupt and will allow us
to easily switch to a method that alternates expansion of forward and reverse
searches (if that proves to be effective).

* Increment n outside of interrupt check (in case we are using without an
interrupt method set and want to alternate searches).