no need for special case breaks boolean
to me it looked like this special case was added to allow the programmer to ask for a numeric type but allow conversions to happen. so like i want int but the value in json was 3.1415. the problem was that if you asked for bool, boo is convertable to double but rapidjson::value::isnumeric returns false so all bool values are not seen