In UpdatePredictedSpeeds there was a check for presence of predicted speed profiles - (#1513)
if none were present the method returned and no updates to the tile were made. This is
incorrect since there may have been free flow or predicted speeds added to the directed
edges. Removing the check and adding a comment so we don't try to over optimize this
method again.