Adjust Matrix threshold (#1555)
* Add a has_traffic flag to TransitionCost and TransitionCostReverse. This defaults to
false. For time dependent rotues where traffic is available at an intersection we want
to lower the influence of transition costs (especially when computing estimated time
to traverse the intersection). This change allows future use within all costing models
that need to be influenced by traffic, but for now only time dependent, depart-at
routes are set to use this flag.

* Double the cost threshold. The computed values (based on max distance) do not give enough
padding in cases where costing or speeds are lower than expected.

* Update for matrix threshold fix.

* Penalize Uturns more.

* Revert u-turn cost change.