Remove addition of transition edges to priority queue / adjacency list. Add an ExpandForwardMM (#1572)
method that is called for multimodal isochrones. This uses a similar pattern to other Expand*
methods in thor. However, multimodal methods have several extra variables that need to be
maintained - these now become class member variables to avoid having to pass them through
methods. The ExpandForwardMM seems too long to be a lambda.