Turnlanes (#1634)
* Add stub for a DirectedEdgeExt - extended directed edge attributes.

* Use bit fields for some counts - check limits. For example, nodecount cannot be more than
kMaxGraphId, so use only 21 bits for this field and check for exceeding the limit.

* Update GraphTileHeader tests and no_bin graph tiles

* Reorder some fields in the GraphTiles. Move turnlanes to be after signs (they are both fixed size
structures and both are indexed by edgeId). Add some support for extended directed edge attributes.
This would be 1:1 structures with DirectedEdge and would allow extra, custom attributes.

* Check has_elevation in trip path builder. Edge elevation flags are not longer
needed in hierarchy builder.

* Move turnlanes. Trim graph tile header by using some bit fields. Make bike_network
a flag in directed edge, and a bit mask in EdgeInfo.