Removed the local ferry_factor_ inside bicycle costing. This conflicted with the ferry_factor_ (#1663)
in the base DynamicCost class and led to undefined behavior. Worse case is this factor was
sometimes set to a negative value which led to negative costs for ferry edges. These negative
costs could lead to one search direction in bidirectional halting - leaving the other search
direction expanding for a potentially very long time without connecting to the other search
direction. This uses up lots of memory and causes long running routes and persistent memory
(allocations in edge labels for instance).