Serialize osmdata (#1715)
* Move UniqueNames methods to .h file and remove the Log method (not really needed).

* Use a single UniqueNames object - move refs, exit_to, node names, and node_ref into the
name_offset_map. Rather than using StringMap for node strings, add indexes to OSMNode
(increases size slighty - from 24 bytes to 32 bytes). All these steps lower the number
of separate data members in OSMData, making it easier to write code to serialize the
class in and out of a file. This will allow splitting the parsing and tile building
stages of valhalla_build_tiles.

* Initialize directededges_ext (was uninitialized).

* Add plumbing for serializing OSMData to and from temporary files.

* Fix formatting

* Finished the OSMData write to and read from temporary files. Added an initialize stage to
the pipeline that cleans up the tile directory (creates it if it doesn't exist). Validated
that the size of the OSMData structures matched when written out and read back in. Total
tile size of the resulting data set matches when going through the temporary file path
as opposed to remaining in memory.

* Fix writing and reading lane connectivity. Needed to move strings into the UniqueNames
list and store indexes in the temporary structure/file.

* Add a script to incrementally build tiles, checking for error returns

* cleaned up.

* cleaned up.