Adds unittest of timedep_forward for whitelion u-turn
It does this by adding whitelion_bristol_uk_reversed_oneway.osm.pbf
which is the same as whitelion_bristol_uk.osm.pbf with the exception
of having the oneway on Quay street reversed.

Reversing that oneway makes the forward expanded A-star able to find
a route through detection of a deadend at runtime.