Heading filter fix for node search (#2058)
* Fix logic for shape index on the opposing (other) edge in the node heading
filter. Need to use opposite logic and make sure the tile and edgeinfo for the
opposing edge is used in case it was from a different tile and the forward flag
is the same as the current edge.

* add bug fix

* Formatting fix

* Can still use the original edge info / shape. The forward flag is set based on orientation
of OSM nodes on the original "edge" so opposing edges in adjacent tiles should have opposing
forward flags.