Fixes false positive astar-tests
Some of the astar-tests were false negatives. They passed even though
they result in NoRoute.

Reason is that the expected elapsed time is erranously calculated to be 0,
the same elapsed time computed for a NoRoute, thus resulting in Pass.

The NoRoute in turn seems to come from the fake_tiles not being built
by default, unless MAKE_TEST_TILES is defined. This seems to have been
disabled for a long time since the code behind this define does not compile.

The tile building code had also had some bit-rot from being left out of
compiling for so long.

This commit enables the tile-building code and builds the tiles as part of the test
if needed.