[tests] Use googletest in unit tests
googletest reporistory is used as a git submoudule.
googletest and googlemock are linked to every test.

All unit tests are converted to to use gtest/gmock primiteves
This allowed to make the faliure output more informative without extra messages.

Lots of bugs in tests have been fixed along the way:
- Incorrect check conditions (different flavors or errors)
- Some test functions were there but not registered in test suite

The test that were not in a suite and currently failing are disabled.
Separate issues will be created for them (affects only a couple of test cases).

The original test suite has been completely removed.
test.h is still used for auxiliary goodies.

Existing tests have plenty of examples for:
- differet kinds of assrts/expect calls
- usage of test fixutres
- usage of environment setup/teardown for a suite
- basic exmamples of gmock matchers

This is a combination of 126 commits.

Closes #2130