Update batch.sh to be compatible with macos utils
One of the sed commands in run_route_scripts/batch.sh fails to run on
macos with the following error:

sed: -e: No such file or directory

Unfortunately, I was not able to get to the root of the problem (in part
because I'm unable to run dtrace in a restricted priviledge level on my
mac).

A quick fix, which should work on Linux based systems as well, is to
pass an emoty string to the -i option. This should have the same effect
as not supplying it (which tells sed to make inline changes without
making a backup).

Tested by running batch.sh as follows:
$ ./batch.sh ../test_requests/demo_routes.txt ../build/valhalla.json