Reverts to `kInvalidSecondsOfWeek` for `SetOrigin` in timedep_forward
The tests were failing beause they were either not correctly setting the
date-time on the origin/destination locations or setting it to 23:05
which pretty much matches `kFreeFlowSpeed` while a `kInvalidSecondsOfWeek`
evaluates to `kConstrainedSpeed` which is speed around noon.

Setting the origin/destination datetimes to 13:05 rather than 23:05
and increasing the expected durations solved this so I can go back
to `kInvalidSecondsOfWeek`.