Adds stricter unittest of Isochrones
Yes, hardcoding an expected response from Isochrones is
not ideal but it is at least much better then the current
test that basically checks if Isochrones returned succesfully.

Given the coming refactorings to Isochrones, this seems like
a good first sanity check.

Compare json docs and not serialized strings
as order is not defined once serialized