Rework language files for Transifex translations
This PR introduces a new language source file in YAML format. This format is friendlier for translation purposes because it:
- is easier to read
- allows inline comments
- has the ability to specify aliases so we don't repeat phrases
- is understood by the Transifex translation system

Transifex is a community translation platform (https://www.transifex.com/) which makes it easy for anyone to contribute to these language files. It has a free tier for open source projects that we can use for the Valhalla project. It also has a GitHub integration to automatically pull in changes. This will be much easier for external contributors to use.

It does introduce an extra "build" step for our language files. In this PR the structure of the JSON language files is unchanged, but now we "compile" the YAML into the JSON format that Odin expects. We also now automatically generate example phrases, which means it's no longer necessary for translators to edit examples manually. The build step is an executable python script: build.py.