Restores EdgeStatus as erroneously set as part of complex restriction (#2109)
* Corrects return value for permanent edge in timedep astar

If the inner expand function encounters a permanent edge, it's an edge
that _could_ have been expanded and so the Inner functions should return
true. This is already the case in bidirectional but the timedependent
functions seem to have missed this tweak

* Restores EdgeStatus as erroneously set as part of complex restriction

When a complex turn restriction triggers and prevents further expansion,
earlier edges in the restriction have already been marked as kPermanent.
This leads to early termination of that entire subtree of the graph
expansion as documented in https://github.com/valhalla/valhalla/issues/2103

The change introduced here goes back to the previous edge marked
kPermanent and marks it as kUnreached so that it may be expanded again
when evaluated from a different part of the graph.

Fixes https://github.com/valhalla/valhalla/issues/2103

* test: Enlarges bayfront pbf so that Bidirectional astar unittest can route

* Also enables complex restriction backtracing for regular a-starr

* Don't reset the first restriction edge when backtracking

since there is no point in possibly expanding A more than once
and carries the potential of spawning infinite loops of expansion

* Applies complex restriction reset in bidir-reverse direction too

* Adds test-case for behaviour added in 8c8e415f17dc0

That commit resets EdgeStatus' whenever a complex restriction is
triggered. This test verifies that this reset does not break behaviour
for when a detour occurs _after_ the partial complex restriction.

References https://github.com/valhalla/valhalla/issues/2103#issuecomment-565727968

* Renames kUnreached to the more appropriate kUnreachedOrReset

* style: format code

* nit: Changes list to vec