fixes(traffic-speed): Fixes underflow for traffic speed > 140kph
The apparent underflow happened because of the use of the precomputed
division tables in the costing functions. They were pre-computed using
the assumed max speed of 140kph.

Thus, when faced with speeds larger than that from traffic,
out-of-bounds access occured and garbage data was read.

This change tweaks the max speed to account for higher speeds in traffic
data.

Checks speedfactor out-of-bounds in debug with assert