Adds glog (Google logging) support.
This is a draft PR demonstrating the use of glog's configurable
logging and assertion macros. A few benefits: log-level
configurability, thread-safety, pretty-good (but synchronous
performance), assertion macros, and signal handlers with clean
backtraces (what logbt provides). There are really great capabilities
outlined in the documentation linked below.

Regarding performance: My understanding is that we don't have a lot of
high-throughput logging use cases out of Valhalla (like logging at
usec resolution), but if we do and we are extremely concerned about
blocking, my understanding is that we can extend
`google::base::Logger` with a minimal amount of hassle. Here's an
example of how to do that:
https://github.com/apache/kudu/blob/master/src/kudu/util/async_logger.h#L36.

Documentation:
- https://github.com/google/glog/blob/master/src/glog/logging.h.in#L155x
- https://github.com/google/glog/blob/master/doc/glog.html
- Rendered: https://hpc.nih.gov/development/glog.html

Follow up to:
- https://github.com/valhalla/valhalla/pull/2216

/cc @kevinkreiser @a-kosak-mbx @mandeepsandhu