Matching Serialization Refactor (#2292)
Quoth @kevinkreiser: Basically we were doing a bunch of gymnastics to turn the output of meili into something that thor could use like a normal route path and locations. We had to do the gymnastics because Meili was designed to do one big match, not a bunch of legs or potentially multiple routes where discontinuities occur. Please read the above issue for a wider description about what the goal of the PR is and a breif overview of the implementation.

Fixes #2233

Co-authored-by: Yuzhe Yan <yuzhe.yan@mapbox.com>
Co-authored-by: Kevin Kreiser <kevinkreiser@gmail.com>
Co-authored-by: Buro Mookerji <buro.mookerji@mapbox.com>