Fix bug and provide better error message when Lua barfs.
There was a bug when `lua_pcall` returned something other than success. This caused a log message to be printed, but the function then continued into pulling the results off the Lua stack. If these results didn't exist, it was possible to sometimes get SIGSEGV.

Since the Lua code is part of Valhalla, it makes sense that a bad response from Lua should be treated as a logic error and exceptional condition. However, this wouldn't give a lot of information for debugging, so I've added the OSM ID as a parameter to provide better debuggability.