Add utility functions to Signs class
Added utility functions to Signs that return a list of signs instead of
a string but are functionally similar to ListToString & GetGuideString
APIs.

These APIs can be used in place of the string variants when we want to
control how the individual signs are used.

Added tests for the new APIs.