Remove "leaked" dependence on robin-hood-hashing
Modify thor::CostMatrix to make the target map an opaque type. This
allows us to entirely encapsulate use of RHH in the implementation,
avoiding what otherwise appears to be the only place where use of RHH
leaked into the public interface. In particular, this means that
consumers no longer need to be able to include the RHH headers. (Which
is very fortunate, as we weren't installing them!)