Remove non-prefixed paths from public includes
Remove the paths that include the 'valhalla' subdirectory from the set
of public includes. This will require external users to prefix includes
with 'valhalla', which has always been the intent. Add additional
private includes to avoid (for now, at least) having to change all sorts
of local source files.

This is a second step toward providing a cleaner interface for external
consumers. (The first was not 'leaking' object libraries as interface
link libraries, as was done in a previous commit.)