Hide robin hood (#2455)
* Remove "leaked" dependence on robin-hood-hashing

Modify thor::CostMatrix to make the target map an opaque type. This
allows us to entirely encapsulate use of RHH in the implementation,
avoiding what otherwise appears to be the only place where use of RHH
leaked into the public interface. In particular, this means that
consumers no longer need to be able to include the RHH headers. (Which
is very fortunate, as we weren't installing them!)

* Make idtable.h explicitly private

Move idtable.h into src/ so that it is explicitly a private header.
Also, remove its inclusion from pbfgraphparser.cc, which was not
actually using it. This "fixes" the other potential leak of
robin-hood-hashing. (Note also the previous commit.)

Co-authored-by: Kevin Kreiser <kevinkreiser@gmail.com>