Fix use of prime_server
There are two issues with how we were using prime_server.

First, it has a non-exported dependency on ZeroMQ, which would cause
builds to fail if ZeroMQ is not in a system include directory (and
possibly if prime_server is built as a static library, but I'm ignoring
that for now...). Fix this by explicitly finding ZeroMQ also and adding
its include directories to those of prime_server.

Second, at some point (just over two years ago!), prime_server changed
its include directories from `include` to `include/prime_server`, which
means all of our includes are wrong and are unable to find the headers
if they aren't in a system include path. Fix this by scanning the
include paths for libprime_server and, if any end in `prime_server`,
also adding the parent path to the include paths.