Suppress -Wreturn-type warning
Add something to some code paths that are meant to be unreachable so
that the compiler (when building in release mode, in which the `assert`
that would otherwise prevent these points from being reached, is a
no-op) won't complain about execution falling off the end of the
function (i.e. -Wreturn-type tripping).

This could alternatively be handled by an unreachable annotation, but
that is less portable.