Minor fixes to tests (#2483)
* Remove unneeded 'using namespace' in test

Remove 'using namespace prime_server' from thor_worker test. This does
not appear to be needed, and causes a compile error when building
without services enabled, since prime_server is not used in that case,
and therefore the namespace is not known.

* Suppress -Wreturn-type warning

Add something to some code paths that are meant to be unreachable so
that the compiler (when building in release mode, in which the `assert`
that would otherwise prevent these points from being reached, is a
no-op) won't complain about execution falling off the end of the
function (i.e. -Wreturn-type tripping).

This could alternatively be handled by an unreachable annotation, but
that is less portable.