Implement bikeshare (#2031)
* itermedia commit; tobe squashed

* Implement bike share

* minor fix

* format

* lint

* update for latest sif changes

* more changes from master

* even more compat for master

* fix: pedestrian is projected on bss connection

* add new case handle for bikeshare

* take projected edge's access

* fix crash when bss mode is not initialized

* fix bugs: project on bss connections, incorrect bss maneuvers

* add tests

* format

* minor fix

* fix lint

* add test for rent/return cost/penalty

* update doc

* update test data and test

* Fixes to make batch.sh run correctly

The last sed command in this script was creating a malformed json. It
was stripping the trailing single-quote of the json string but not the
leading one.

Changed the sed expression to remove single quotes from the json string.
Also moved adding the "--config" option to the place where parallel is
invoking valhalla_run_route. Since the arg to this option is fixed for
all runs of parallel, I find it easier to read if placed in the same
place where the executable is being called. This also limits the sed
expression to whats needed.

Tested by running batch.sh as follows:
$ ./batch.sh ../test_requests/demo_routes.txt ../build/valhalla.json

* Fix breakage of batch.sh for Linux

Added back the -C '\|' option so that parallel can pass the -j arg
correctly to valhalla_run_route

* Nit: Fixes typo icp -> ipc

Fix the typo in the configuration file check for the http listening protocol in case of the socket

* Update README.md (#2166)

As we have Switzerlan's maps prepared during the tutorial I would like to update the locations in the curl according to the proper coordinates. Otherwise, it's disappointing having incorrect curl which does work at the test step of the tutorial.

* update test

* update valhalla_build_config

* fix after merge master

* fix after merge

* test file clean up

* fix after clang-tidy

* Update narrativebuilder.cc

* Temporary fix to bssbuilder

Temporary - skip BSS nodes that connect to an edge that crosses tile boundaries. These caused crashes in HierarchyBuilder and invalid opposing edges in GraphValidator.

* formatting

* formatting

* Update dynamiccost.h

remove useless blank line

* update test config for astar_bss

* stop using hierarchy limits

* minor fix on test

* minor fix on test

* update test config

* fix compilation

* fix one test

* update test because of Obvious maneuvers

* format

Co-authored-by: Kevin Kreiser <kevinkreiser@gmail.com>
Co-authored-by: Mandeep Sandhu <mandeep.sandhu@mapbox.com>
Co-authored-by: Fedor Shiriaev <theodor.shishi@gmail.com>
Co-authored-by: David Nesbitt <dnesbitt61@gmail.com>