valhalla_build_tiles: Add build manifest and separate edge construction stage (#2511)
* This PR refactors `valhalla_build_tiles` to make it easier to partition and parallelize tile
construction on a batch processing system. To do so, it pulls out edge construction into a new
`constructedges` phase. `constructedges` constructs edges on a global dataset, and also emits a
manifest file of Valhalla tile IDs that can be partitioned and be used to process in parallel during
the build phase. For more about the tile manifest, see comment in `mjolnir/graphbuilder.h`.

This update is intended _not_ to break downstream users of these tools. For example,
`valhalla_build_tiles --start build --end build ...` will create a tile manifest if one does not
already exist in the expected location.

/cc @gknisely @kevinkreiser @merkispavel

* Update tests.

* Updating tests.

* Move manifest into util header

* Test coverage

* Unit test coverage.

* Updated tests

* Actual round trip test.

* Test tile build.

* Cleanup and fix tests.

* Fix formatting issue.

* Code review feedback.

* Future TODO comment to make constructedges fully required.

* Reduce test running parallelism in build-release-x86.

* Revert "Reduce test running parallelism in build-release-x86."

This reverts commit 82795154af6be70293a8d59dea00e1270e1d5fb4.

* Update util_mjolnir.cc

* Update graphbuilder.cc

Co-authored-by: Pavel Merkis <pavel.merkis@mapbox.com>
Co-authored-by: Kevin Kreiser <kevinkreiser@gmail.com>