Penalize closed edges when routing through them (#2964)
Penalize closed edges by applying a "closure_factor" to the closed edge's cost.
This does not change the time to traverse a closed edge, but only its cost.

This ensures we exit out of closures asap. The closure penalty factor can also be
set at request-time in the costing options.