nit(rename): Renames the encoded live speed properties
to make it clear that the raw values are usually not what you want.
They are not measured in kph, but have on several occasions been
used in that way due to the non-obvious naming.

This rename adds `encoded_` to the speed properties to make it
clearer that the value within may not be what you actually want.

There already are getter-functions defined which should be preferred
in most use-cases