Temporarily disable ciignore functionality
We dont want to skip a skippable change, if the previous build is
running (we want the entire CI run to happen again, since the results
of the previous run were not ready).

This requires querying circle CI from "needs_ci_run" and then based
on if a previous build is running or not skip CI. In order not to allow
comitters to accidentally skip CI, this commit disables the skip
feature till we add support for the above issue.