Add RelaxHierarchyLimits with pre-defined algorithm-based factors
Single place of definition of the hard-coded factors which apparently
are re-used in at least two locations.

This also fixes a inconsistency or bug/regression that was introduced around  two commits
- https://github.com/valhalla/valhalla/commit/1b150f7153a41de673acf88a5c22ac7709e91c1a
- https://github.com/valhalla/valhalla/commit/9b646089cf6db833ee712e22e988885ac3930487
which pass a boolean to RelaxHierarchyLimits,  instead of float-point factor.
which was later cultivated unnoticed in PR https://github.com/valhalla/valhalla/pull/2719

Maintaining the factors in single place should help prevent such situations.