Overflow in sequence (#3303)
* Overflow in sequence

priority_queue holds indexes to the elements. And after node insert to the output sequence index is incremented and put back to the priority_queue. When we have more than INT_MAX elements it overflows and drops part of points.

* Changelog

* clang-format

* OSMNode validates value after conversion to fixed point.

Otherwise it could drop valid values in (anti)meridian nodes.

* abs is useless for unsigned values.

* Fixed undefined behavior

We should check value interval before we try to set Infinity into the integer.

Co-authored-by: Evgen Bodunov <evgen@getyourmap.com>