Do not create statsd client in workers if it is not configured
Right now the statsd client is created all the time.
If there is no configuration, the defaults are used.
However, under some condigitions we noticed unwanted side effects.

For example, the use of getaddrinfo during initialization may cause DNS requests under the hood.
In case of the bogus network on the start, it may cause uncontrollable hangs.

The idea here is not to pay for what is not used.