Use std::multimap for admin and timezone polygon results (#3427)
* Change admin polys and tz polys to be stored in std::multimap instead of
std::unordered_multimap. This preserves the order so that polygons that
are the result of state queries are tested first in GetMultiPolyId. With
unordered_multipmap, the order is indeterminant and the country polygon
can be checked first. This is sub-optimal since they are generally much
larger, and often would then need to also check the state polygons as well.
Another optimization is to avoid the country query if there is only 1
multipolygon resulting from the state query AND the bounding box is
entirely contained within the state polygon. In this case it is safe to
set the tile_within_one_admin flag to true which will avoid calling boost
geometry covered_by for each node lat,lng within the tile.

* Revert change that avoided country query if bounding box was entirely within the
single state query result. This change led to tile size changes which may be ok
but was not verified. The change without this still leads to substantial performance
improvement.

* Formatting.

* Add admin poly change

* remove unused type alias

* Remove unused code

Co-authored-by: Kevin Kreiser <kevinkreiser@gmail.com>