Updates to country specific access overrides (#3578)
* Updates to country specific access. In particular, add kBicycle access to trunk
roads in Romania. Add a couple countries that did not previously have any
override logic. Also, not some that are in the OSM Wiki but do not have an entry
here since there are no actual overrides at this time.

* Update Netherlands so bicycle is allowed by default for trunk (and trunk_link). If motorroad
is set then the code in country_access should disable bicycle access by default (unless there
is a motorroad override present).

* Update change log