Fix Build on MacOS (#4185)
* Remove unused variable

Apple clang generating this error:
 error: variable 'found' set but not used [-Werror,-Wunused-but-set-variable]

* Replace set with unordered_set

std::set is not defined in libc++ on MacOS XCode SDK 12 and is defined with a different type signature in XCode 13. There doesn't seem to be any need for the set to be ordered so this fixes the compilation errors

* Add maybe_unused attribute to found variable

This was providing a similar error as before:
 error: variable 'found' set but not used [-Werror,-Wunused-but-set-variable]
As the variable is only used when the logging level is set to trace.