Ferry reclassification fix when multiple edges exist between a node pair (#4361)
* When adding a node label, add the edge's wayindex. When recovering the
shortest path, make sure the wayindex of the edge connecting the node
with its predecessor node matches the wayindex. This ensures the correct
edge on the shortest path is marked for reclassificiation. Prior to
this, multiple edges could become reclassified when there is more than
1 edge between a node pair.

* Change log

* Fix use of way_index when recovering shortest path.

* Add a test to make sure only the shorter edge between a node pair is
reclassified. Tested this against master and it fails. Succeeds on
the branch.

* formatting fix