Make test_valhalla_build_extract.py more robust to changes.
Testing with hardcoded offsets and file sizes is very brittle when anything changes for any reason,
and leads to the test just being modified by putting in whatever numbers will make it pass (i.e. for
the latest case see: https://github.com/valhalla/valhalla/commit/678834f0add5fa54347cad59abc0818910f32b26)

Instead restructure so we're testing what we actually care about - i.e. that the offsets are in the
right places based on the tile sizes, that the traffic tiles have space for the same number of edges
as in the main tiles etc.