Add Rust polyline decoding example
This adds a Rust implementation to the documentation on decoding a route
shape.

While other implementations only add the function itself, this patch
also adds a simple test to demonstrate working behavior. A language
implementation might not be the correct place for this, but providing
examples for correct behavior seems useful as a reference for people
looking to implement this themselves.